Users CRUD

+++ src/main/java/com/book/find/controllers/UsersController.java
@@ -1,5 +1,6 @@
 package com.book.find.controllers;
 
+import com.book.find.dto.SuperAdminDTO;
 import com.book.find.models.Auth.User;
 import com.book.find.repos.UserRepo;
 import org.springframework.stereotype.Controller;
@@ -8,8 +9,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.servlet.ModelAndView;
 
 import java.io.IOException;
-import java.security.Principal;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
@@ -23,15 +22,16 @@ public class UsersController {
     }
 
     @GetMapping("/users")
-    public String main(Map<String, Object> model){
-        List<User> users = userRepo.findAll();
-        model.put("films", users);
+    public String main(Map<String, Object> model) {
+        model.put("categories", userRepo.findAll());
         return "users";
     }
 
     @GetMapping("/deleteUser/{id}")
     public ModelAndView delete(@PathVariable UUID id) throws IOException {
         User user = userRepo.findById(id);
+        if(user.getUsername().equals(SuperAdminDTO.username))
+            return new ModelAndView("redirect:/users");
         userRepo.delete(user);
         return new ModelAndView("redirect:/users");
     }
